Use Java 17

+++ java/src/main/java/io/cucumber/prettyformatter/PrettyReportData.java
@@ -211,7 +211,7 @@ final class PrettyReportData {
         return query.findPickleBy(testCaseStarted);
     }
 
-    Optional<Long> findLineOf(Pickle pickle) {
+    Optional<Integer> findLineOf(Pickle pickle) {
         return query.findLocationOf(pickle).map(Location::getLine);
     }
